+++ soske/comms/__init__.py
@@ -3,6 +3,7 @@ Package definition for 'soske.comms'.
 """
 
 # Explicit imports to activate group commands.
+from soske.comms.delete import delete
 from soske.comms.get import get
 from soske.comms.list_ import list_
 from soske.comms.set_ import set_