+++ changes.md
@@ -1,11 +1,13 @@
 # Changes
 
 ## [Unreleased]
+### Added
+- New commands: `delete`.
 
 ## [v0.1.0] - 2023-03-01
 ### Added
 - Basic codebase and tests.
-- New commands: `get`, `lst`, `set`.
+- New commands: `get`, `list`, `set`.
 
 [Unreleased]: https://github.com/wirehaiku/soske/tree/main
 [v0.1.0]:     https://github.com/wirehaiku/soske/tree/v0.1.0